refactor(CrewRepository): 클래스 이름 변경, repository 패키지 삭제, domain 패키지로 재이동
